Clean up loading functions
