+++ microhapulator/tests/test_filter.py
@@ -11,11 +11,10 @@
 # -------------------------------------------------------------------------------------------------
 
 import microhapulator
-from microhapulator import load_marker_thresholds
 import microhapulator.api as mhapi
 from microhapulator.profile import TypingResult
 from microhapulator.tests import data_file
-import pandas as pd
+from microhapulator.thresholds import ThresholdIndex
 import pytest
 
 
@@ -23,7 +22,7 @@ def test_filter_simple():
     bam = data_file("pashtun-sim/aligned-reads.bam")
     tsv = data_file("pashtun-sim/tiny-panel.tsv")
     observed = mhapi.type(bam, tsv)
-    thresholds = load_marker_thresholds(observed.markers(), global_static=10, global_dynamic=0.05)
+    thresholds = ThresholdIndex.load(global_static=10, global_dynamic=0.05)
     observed.filter(thresholds)
     expected = TypingResult(fromfile=data_file("pashtun-sim/test-output.json"))
     assert observed == expected
@@ -31,8 +30,8 @@ def test_filter_simple():
 
 def test_filter_config_file():
     result = TypingResult(fromfile=data_file("prof/deep-raw.json"))
-    thresholds = load_marker_thresholds(
-        result.markers(), configfile=data_file("filters.csv"), global_static=5, global_dynamic=0.02
+    thresholds = ThresholdIndex.load(
+        configfile=data_file("filters.csv"), global_static=5, global_dynamic=0.02
     )
     result.filter(thresholds)
     assert len(result.haplotypes("mh01XYZ-1")) == 8
@@ -41,25 +40,17 @@ def test_filter_config_file():
 
 
 def test_filter_missing_column():
-    result = TypingResult(fromfile=data_file("prof/deep-raw.json"))
     with pytest.raises(ValueError, match=r"filter config file missing column\(s\): Static"):
-        thresholds = load_marker_thresholds(
-            result.markers(),
-            configfile=data_file("filters-missing.csv"),
-            global_static=5,
-            global_dynamic=0.02,
+        ThresholdIndex.load(
+            configfile=data_file("filters-missing.csv"), global_static=5, global_dynamic=0.02
         )
 
 
 def test_filter_dupl_marker():
-    result = TypingResult(fromfile=data_file("prof/deep-raw.json"))
     message = "filter config file contains duplicate entries for some markers"
     with pytest.raises(ValueError, match=message):
-        thresholds = load_marker_thresholds(
-            result.markers(),
-            configfile=data_file("filters-redundant.csv"),
-            global_static=5,
-            global_dynamic=0.02,
+        ThresholdIndex.load(
+            configfile=data_file("filters-redundant.csv"), global_static=5, global_dynamic=0.02
         )
